Move hord storage package to storage/kv
