+++ examples/storage/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/calvinmclean/babyapi"
-	"github.com/calvinmclean/babyapi/storage"
+	"github.com/calvinmclean/babyapi/storage/kv"
 	"github.com/madflojo/hord/drivers/hashmap"
 )
 
@@ -18,14 +18,14 @@ func main() {
 		func() *User { return &User{} },
 	)
 
-	db, err := storage.NewFileDB(hashmap.Config{
+	db, err := kv.NewFileDB(hashmap.Config{
 		Filename: "storage.json",
 	})
 	if err != nil {
 		panic(err)
 	}
 
-	api.SetStorage(storage.NewClient[*User](db, "User"))
+	api.SetStorage(kv.NewClient[*User](db, "User"))
 
 	api.RunCLI()
 }
