+++ storage/kv/client_test.go
@@ -1,4 +1,4 @@
-package storage
+package kv
 
 import (
 	"context"
